Ensure RoPE returns same dtype as input
Cast RoPE output to the same dtype as the input tensor X to avoid dtype mismatches (e.g., during mixed-precision) and potential downstream errors.

+++ Preprocess/pos_embed.py
@@ -49,4 +49,4 @@ class RoPE(nn.Module):
         x1_rot = x1 * cos - x2 * sin
         x2_rot = x1 * sin + x2 * cos
         
-        return torch.cat([x1_rot, x2_rot], dim=-1)
+        return torch.cat([x1_rot, x2_rot], dim=-1).type_as(X)
